moved gated mixture to `models` module
